More Caching, Datatypes fixes (Thanks JDBC) and Measure Diagrams

See merge request team-symphonie/hoots!33

+++ src/hoot/model/entities/Hoot.java
@@ -20,7 +20,7 @@ public abstract class Hoot
 
     public HootTags tags;
 
-    public Map<Interaction, Integer> reactionCount = new HashMap<>();
+    public Map<Interaction, Long> reactionCount = new HashMap<>();
 
     public Hoot()
     {
